Created basic logic to send and receive an object w/ a functionality test.
Trying to write this messages has been hard... the names for various things needs to be completely redone.

Created new ArbiterException: MismatchedObject.

Created a new map to map the raw representation of each type, to the ENUM type.

The test uses 2 threads to send and receive data over a socket. It compares the received data to a newly constructed object.