Refactored the type maps. Broke NetworkElements into 2 types.
Created the NetworkElement class to hold information about each 'element'/field.

Created a single function to return a list of NetworkElements of a given class

Created the NetworkObjectStructure class to hold the basic elements and advanced elements.

Basic elements can be represented statically i.e. constant size, not null.
Advanced elements will be arrays, strings, etc. and will be at the end of the object in their own order.